feature: add auth code

See merge request s07-ai-speech-sub2/S07P22B303!7

+++ backend/app/routers/recipes.py
@@ -1,58 +1,75 @@
-from fastapi import APIRouter, Response, status
+from fastapi import APIRouter, Response, status, Header, Depends
 from typing import Union
 
 from starlette.responses import JSONResponse
-from tortoise.expressions import Subquery
+
+from app.routers.accounts import oauth2_scheme, get_current_user
 
 from app.models.recipes import Recipe, Tag, Ingredient, RecipeComment, LikeRecipe, RecipeIngredient
 from app.models.accounts import User
 
 from app.schemas.recipes import RecipeCreateForm, TagForm, IngredientForm, IngredientRecipeForm, RecipeCommentForm, \
-    LikeRecipeForm, RecipeCommentList, RecipeForm
+    RecipeCommentList
 from app.schemas.common import *
 
 router = APIRouter(prefix="/recipe", tags=["recipe"])
 
 
+
+
 @router.post("/tag", description="레시피 태그 생성", response_model=CommonResponse)
-async def create_tag(req: TagForm):
-    new_tag = await Tag.get_or_create(**req.dict())
+async def create_tag(req: TagForm, token: str = Depends(oauth2_scheme)):
+    user = await get_current_user(token)
+    if user is not None and user.is_admin is True:
+        await Tag.get_or_create(**req.dict())
+    else:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
     return CommonResponse()
 
 
 @router.post("/ingredient", description="레시피 재료 생성", response_model=CommonResponse)
-async def create_ingredient(req: IngredientForm):
-    await Ingredient.get_or_create(**req.dict())
+async def create_ingredient(req: IngredientForm, token: str = Depends(oauth2_scheme)):
+    user = await get_current_user(token)
+    if user is not None and user.is_admin is True:
+        await Ingredient.get_or_create(**req.dict())
+    else:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
     return CommonResponse()
 
 
 @router.post("/", description="레시피 작성", response_model=CommonResponse, status_code=201)
-async def create_recipe(req: RecipeCreateForm):
+async def create_recipe(req: RecipeCreateForm, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
-    recipe = await Recipe.create(user_id=req.user_id, title=req.title, content=req.content, img_url=req.img_url)
-    for ingredient in req.ingredients:
-        new_ingredient, created= await Ingredient.get_or_create(name=ingredient.name)
-        await RecipeIngredient.create(recipe=recipe, ingredient=new_ingredient, amount=ingredient.amount)
-    await recipe.tags.add(*[tag for tag in await Tag.filter(id__in=req.tags)])
+    user = await get_current_user(token)
+    if user is not None:
+        recipe = await Recipe.create(user_id=user.id, title=req.title, content=req.content, img_url=req.img_url)
+        for ingredient in req.ingredients:
+            new_ingredient, created = await Ingredient.get_or_create(name=ingredient.name)
+            await RecipeIngredient.create(recipe=recipe, ingredient=new_ingredient, amount=ingredient.amount)
+        await recipe.tags.add(*[tag for tag in await Tag.filter(id__in=req.tags)])
+    else:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
     return CommonResponse()
 
 
 @router.get("/detail/{recipe_id}", description="레시피 상세", response_model=ObjectResponse)
 async def recipe_detail(recipe_id: int):
-    recipe = await Recipe.get(id=recipe_id)
-    ingredients = [
-        IngredientRecipeForm(**dict(await ingredient.ingredient), amount=ingredient.amount)
-        for ingredient in await RecipeIngredient.filter(recipe_id=recipe_id)
-    ]
-    data = {
-        "recipe": recipe,
-        "nickname": (await recipe.user).nickname,
-        "tags": await recipe.tags.all(),
-        "ingredients": ingredients,
-        "like_users": await recipe.like_users.all().values("id", "nickname"),
-        # comments = await RecipeComment.filter(recipe_id=recipe_id)
-    }
-
+    recipe = await Recipe.get_or_none(id=recipe_id)
+    if recipe is None:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
+    else:
+        ingredients = [
+            IngredientRecipeForm(**dict(await ingredient.ingredient), amount=ingredient.amount)
+            for ingredient in await RecipeIngredient.filter(recipe_id=recipe_id)
+        ]
+        data = {
+            "recipe": recipe,
+            "nickname": (await recipe.user).nickname,
+            "tags": await recipe.tags.all(),
+            "ingredients": ingredients,
+            "like_users": await recipe.like_users.all().values("id", "nickname"),
+            # comments = await RecipeComment.filter(recipe_id=recipe_id)
+        }
     return ObjectResponse(data=data)
 
 
@@ -64,20 +81,25 @@ async def get_comment_list(recipe_id: int):
 
 
 @router.post("/comment/{recipe_id}", description="레시피 댓글 작성", response_model=CommonResponse, status_code=201)
-async def create_comment(recipe_id: int, req: RecipeCommentForm):
+async def create_comment(recipe_id: int, req: RecipeCommentForm, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
-    await RecipeComment.create(recipe_id=recipe_id, **req.dict())
+    user = await get_current_user(token)
+    if user is not None:
+        await RecipeComment.create(recipe_id=recipe_id, user_id=user.id, **req.dict())
+    else:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
     return CommonResponse()
 
 
 @router.put("/comment/{comment_id}", description="레시피 댓글 수정", response_model=CommonResponse)
-async def edit_comment(comment_id: int, req: RecipeCommentForm):
+async def edit_comment(comment_id: int, req: RecipeCommentForm, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
+    user = await get_current_user(token)
     comment = await RecipeComment.get_or_none(id=comment_id)
     if comment is None:
         return JSONResponse(status_code=404, content=CommonFailedResponse(detail="없는 댓글입니다.").dict())
     else:
-        if req.user_id == comment.user_id:
+        if user.id == comment.user_id:
             comment.content = req.content
             await comment.save()
         else:
@@ -86,18 +108,23 @@ async def edit_comment(comment_id: int, req: RecipeCommentForm):
 
 
 @router.delete("/comment/{comment_id}", description="레시피 댓글 삭제", response_model=CommonResponse)
-async def delete_comment(comment_id: int):
+async def delete_comment(comment_id: int, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
-    await RecipeComment.filter(comment_id=comment_id).delete()
+    user = await get_current_user(token)
+    comment = await RecipeComment.get_or_none(id=comment_id)
+    if user.id == comment.user_id:
+        await RecipeComment.filter(id=comment_id).delete()
+    else:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
     return CommonResponse()
 
 
 @router.post("/like/{recipe_id}", description="레시피 좋아요", response_model=ObjectResponse, status_code=201)
-async def like_recipe(recipe_id: int, req: LikeRecipeForm, response: Response):
+async def like_recipe(recipe_id: int, response: Response, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
     recipe = await Recipe.get_or_none(id=recipe_id)
-    user = await User.get_or_none(id=req.user_id)
-    like_recipe_list = await LikeRecipe.filter(user_id=req.user_id, recipe_id=recipe_id)
+    user = await get_current_user(token)
+    like_recipe_list = await LikeRecipe.filter(user_id=user.id, recipe_id=recipe_id)
     if recipe is None:
         return JSONResponse(status_code=404, content=CommonFailedResponse(detail="추천할 레시피가 없습니다.").dict())
     elif user is None:
@@ -108,11 +135,11 @@ async def like_recipe(recipe_id: int, req: LikeRecipeForm, response: Response):
             "like": True
         }
         if not like_recipe_list:
-            await LikeRecipe.create(user_id=req.user_id, recipe_id=recipe_id)
+            await LikeRecipe.create(user_id=user.id, recipe_id=recipe_id)
             data["method"] = 'post'
             data["like"] = True
         else:
-            await LikeRecipe.filter(user_id=req.user_id, recipe_id=recipe_id).delete()
+            await LikeRecipe.filter(user_id=user.id, recipe_id=recipe_id).delete()
             data["method"] = 'delete'
             data["like"] = False
             response.status_code = 200
@@ -120,20 +147,27 @@ async def like_recipe(recipe_id: int, req: LikeRecipeForm, response: Response):
 
 
 @router.put("/{recipe_id}", description="레시피 수정", response_model=SingleResponse)
-async def edit_recipe(recipe_id: int, req: RecipeCreateForm):
+async def edit_recipe(recipe_id: int, req: RecipeCreateForm, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
+    user = await get_current_user(token)
     recipe = await Recipe.get_or_none(id=recipe_id)
     if recipe is None:
         return JSONResponse(status_code=404, content=CommonFailedResponse(detail="없는 레시피입니다.").dict())
     else:
-        await recipe.update(**req.dict())
+        if recipe.user_id == user.id:
+            await recipe.update(**req.dict())
         return SingleResponse(id=recipe_id)
 
 
-@router.post("/{recipe_id}", description="레시피 삭제", response_model=CommonResponse)
-async def delete_recipe(recipe_id: int):
+@router.delete("/{recipe_id}", description="레시피 삭제", response_model=CommonResponse)
+async def delete_recipe(recipe_id: int, token: str = Depends(oauth2_scheme)):
     # 유저 인증 로직
-    await Recipe.filter(id=recipe_id).delete()
+    user = await get_current_user(token)
+    recipe = await Recipe.get_or_none(id=recipe_id)
+    if user is not None and recipe is not None and user.id == recipe.user_id:
+        await Recipe.filter(id=recipe_id).delete()
+    else:
+        return JSONResponse(status_code=404, content=CommonFailedResponse(detail="").dict())
     return CommonResponse()
